requirements.txt
customtkinter
psutil
Pillow
pyinstaller
pystray
